feat: visual studio solution

+++ src/client/animatedtext.cpp
@@ -93,7 +93,7 @@ bool AnimatedText::merge(const AnimatedTextPtr& other)
     if(other->getCachedText().getFont() != m_cachedText.getFont())
         return false;
 
-    if(m_animationTimer.ticksElapsed() > Otc::ANIMATED_TEXT_DURATION / 2.5)
+    if(m_animationTimer.ticksElapsed() > static_cast<float>(Otc::ANIMATED_TEXT_DURATION) / 2.5)
         return false;
 
     try {
